Migrate victory-shared-events to TypeScript
