--- packages/victory-shared-events/src/index.js
@@ -1,1 +1,0 @@
-export { default as VictorySharedEvents } from "./victory-shared-events";
