+++ packages/victory-shared-events/src/index.ts
@@ -1,0 +1,1 @@
+export { VictorySharedEvents } from "./victory-shared-events";
